refactor: ATLAS-431: Separate MAC import and dictionary lookup entry points
