+++ Atlas.MultipleAlleleCodeDictionary.Test.Integration/DependencyInjection/ServiceConfiguration.cs
@@ -6,7 +6,8 @@ using Atlas.Common.Notifications;
 using Atlas.Common.Utils.Extensions;
 using Atlas.MultipleAlleleCodeDictionary.AzureStorage.Repositories;
 using Atlas.MultipleAlleleCodeDictionary.ExternalInterface;
-using Atlas.MultipleAlleleCodeDictionary.MacImportServices.SourceData;
+using Atlas.MultipleAlleleCodeDictionary.ExternalInterface.DependencyInjection;
+using Atlas.MultipleAlleleCodeDictionary.Services.MacImportServices.SourceData;
 using Atlas.MultipleAlleleCodeDictionary.Settings;
 using Atlas.MultipleAlleleCodeDictionary.Test.Integration.IntegrationTests.TestHelpers;
 using Atlas.MultipleAlleleCodeDictionary.Test.Integration.Repositories;
@@ -28,6 +29,11 @@ namespace Atlas.MultipleAlleleCodeDictionary.Test.Integration.DependencyInjectio
                 OptionsReaderFor<ApplicationInsightsSettings>(),
                 OptionsReaderFor<MacDictionarySettings>()
             );
+            services.RegisterMacImport(
+                OptionsReaderFor<ApplicationInsightsSettings>(),
+                OptionsReaderFor<MacDictionarySettings>(),
+                OptionsReaderFor<MacImportSettings>()
+            );
             services.SetUpIntegrationTestServices();
             services.SetUpMockServices();
             return services.BuildServiceProvider();
@@ -37,6 +43,7 @@ namespace Atlas.MultipleAlleleCodeDictionary.Test.Integration.DependencyInjectio
         {
             services.RegisterAsOptions<ApplicationInsightsSettings>("ApplicationInsights");
             services.RegisterAsOptions<MacDictionarySettings>("MacDictionary");
+            services.RegisterAsOptions<MacImportSettings>("MacDictionary:Import");
         }
 
         private static void SetUpConfiguration(this IServiceCollection services)
@@ -60,16 +67,16 @@ namespace Atlas.MultipleAlleleCodeDictionary.Test.Integration.DependencyInjectio
             services.AddScoped(sp => Substitute.For<INotificationSender>());
         }
 
-        public static void SetUpMacDictionaryWithFileBackedRepository(this IServiceCollection services,
+        public static void SetUpMacDictionaryWithFileBackedRepository(
+            this IServiceCollection services,
             Func<IServiceProvider, ApplicationInsightsSettings> fetchApplicationInsightsSettings,
             Func<IServiceProvider, MacDictionarySettings> fetchMacDictionarySettings)
         {
-            
             services.RegisterMacDictionary(
                 fetchApplicationInsightsSettings,
                 fetchMacDictionarySettings
             );
-            
+
             services.RegisterLifeTimeScopedCacheTypes();
             services.AddScoped<IMacRepository, FileBackedMacDictionaryRepository>();
         }
