+++ Atlas.MultipleAlleleCodeDictionary.Test.Integration/appsettings.json
@@ -4,6 +4,9 @@
   },
   "MacDictionary": {
     "AzureStorageConnectionString": "UseDevelopmentStorage=true",
-    "TableName": "TestAtlasMultipleAlleleCodes"
+    "TableName": "TestAtlasMultipleAlleleCodes",
+    "Import": {
+      "MacSourceUrl": "not-used-in-tests"
+    }
   }
 }
