+++ Atlas.MultipleAlleleCodeDictionary/Atlas.MultipleAlleleCodeDictionary.csproj
@@ -19,9 +19,5 @@
     <InternalsVisibleTo Include="Atlas.MultipleAlleleCodeDictionary.Test.Integration" />
     <InternalsVisibleTo Include="DynamicProxyGenAssembly2" />
   </ItemGroup>
-
-  <ItemGroup>
-    <Folder Include="ExternalInterface\DependencyInjection" />
-  </ItemGroup>
   
 </Project>
