--- Atlas.MultipleAlleleCodeDictionary/ExternalInterface/ServiceConfiguration.cs
@@ -1,50 +1,0 @@
-﻿using System;
-using Atlas.Common.ApplicationInsights;
-using Atlas.Common.Caching;
-using Atlas.Common.Utils.Extensions;
-using Atlas.MultipleAlleleCodeDictionary.AzureStorage.Repositories;
-using Atlas.MultipleAlleleCodeDictionary.MacCacheService;
-using Atlas.MultipleAlleleCodeDictionary.MacImportServices;
-using Atlas.MultipleAlleleCodeDictionary.MacImportServices.SourceData;
-using Atlas.MultipleAlleleCodeDictionary.Services;
-using Atlas.MultipleAlleleCodeDictionary.Settings;
-using Atlas.MultipleAlleleCodeDictionary.utils;
-using Microsoft.Extensions.DependencyInjection;
-
-namespace Atlas.MultipleAlleleCodeDictionary.ExternalInterface
-{
-    public static class ServiceConfiguration
-    {
-        public static void RegisterMacDictionary(
-            this IServiceCollection services,
-            Func<IServiceProvider, ApplicationInsightsSettings> fetchApplicationInsightsSettings,
-            Func<IServiceProvider, MacDictionarySettings> fetchMacDictionarySettings)
-        {
-            services.RegisterSettings(fetchApplicationInsightsSettings, fetchMacDictionarySettings);
-            services.RegisterServices();
-            services.RegisterAtlasLogger(fetchApplicationInsightsSettings);
-            services.RegisterLifeTimeScopedCacheTypes();
-        }
-
-        private static void RegisterSettings(
-            this IServiceCollection services,
-            Func<IServiceProvider, ApplicationInsightsSettings> fetchApplicationInsightsSettings,
-            Func<IServiceProvider, MacDictionarySettings> fetchMacDictionarySettings)
-        {
-            
-            services.MakeSettingsAvailableForUse(fetchApplicationInsightsSettings);
-            services.MakeSettingsAvailableForUse(fetchMacDictionarySettings);
-        }
-        
-        private static void RegisterServices(this IServiceCollection services)
-        {
-            services.AddScoped<IMacRepository, MacRepository>();
-            services.AddScoped<IMacParser, MacLineParser>();
-            services.AddScoped<IMacImporter, MacImporter>();
-            services.AddScoped<IMacCodeDownloader, MacCodeDownloader>();
-            services.AddScoped<IMacCacheService, MacCacheService.MacCacheService>();
-            services.AddScoped<IMacDictionary, MacDictionary>();
-            services.AddScoped<IMacExpander, MacExpander>();
-        }
-    }
-}
