+++ Atlas.MultipleAlleleCodeDictionary/Settings/MacImportSettings.cs
@@ -1,0 +1,10 @@
+﻿namespace Atlas.MultipleAlleleCodeDictionary.Settings
+{
+    /// <summary>
+    /// Settings only needed for import, but not reading, of the MAC Dictionary.
+    /// </summary>
+    public class MacImportSettings
+    {
+        public string MacSourceUrl { get; set; }
+    }
+}
